Revert "feat(crates): Updated Dependencies & changed ids from i32 to i64"
This reverts commit 49c7d03d404ddc05af9d2ab019491b17d25ce961.

+++ src/entity/championship.rs
@@ -4,8 +4,8 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, FromRow, Clone)]
 pub struct Championship {
-    pub id: i64,
-    pub user_id: i64,
+    pub id: i32,
+    pub user_id: i32,
     pub created_at: DateTime<Utc>,
     pub name: String,
     pub port: i16,
